Renamed _acutator.py to actuator.py.
